refactor(studio): add a generic way of making context by wrapping react utils
