+++ apps/studio/src/data/context/app.tsx
@@ -1,6 +1,6 @@
 import useDisclosure from "~/utils/hooks/useDisclosure";
-import { createContext, ReactNode, useContext } from "react";
-import { isUndefined } from "@biolnk/utils";
+import { makeContext } from "~/utils/makeContext";
+import { ReactNode } from "react";
 
 export type AppContextState = {
   openAddLinkDialog: () => void;
@@ -12,9 +12,10 @@ export type AppProviderProps = {
   children: ReactNode;
 };
 
-const AppContext = createContext<AppContextState | undefined>(undefined);
+const [AppContext, Provider, useAppContext] =
+  makeContext<AppContextState>("AppContext");
 
-export function AppContextProvider({ children }: AppProviderProps) {
+function AppContextProvider({ children }: AppProviderProps) {
   const { isOpen, onOpen, onClose } = useDisclosure();
 
   const value: AppContextState = {
@@ -23,15 +24,7 @@ export function AppContextProvider({ children }: AppProviderProps) {
     isAddLinkDialogOpen: isOpen,
   };
 
-  return <AppContext.Provider value={value}>{children}</AppContext.Provider>;
+  return <Provider value={value}>{children}</Provider>;
 }
 
-export function useAppContext() {
-  const ctx = useContext(AppContext);
-
-  if (isUndefined(ctx)) {
-    throw new Error("AppContext must be used within the 'AppContextProvider'");
-  }
-
-  return ctx;
-}
+export { AppContext, AppContextProvider, useAppContext };
