+++ apps/studio/src/lib/supabase/index.ts
@@ -1,3 +1,2 @@
 export * from "./client";
 export * from "./provider";
-export * from "./context";
